\name{ComparisonCR}
\docType{package}
\alias{ComparisonCR-package}
\title{Comparison of two treatments in the presence of competing risks}
\description{Functions for comparison of two treatments in the presence of competing risks, including statistical inference for overall tests and arbitrary tests based on the area between the CIF curves (Lyu, 2020), and for tesing at a fixed point (Chen, 2020).
}
\author{Jingjing Lyu, Zheng Chen, Jinbao Chen, Yawen Hou.
maintainer: Zheng Chen \email{zheng-chen@hotmail.com}}
\references{
[1] Lyu J, Chen J, Hou Y, Chen Z. Comparison of two treatments in the presence of competing risks. Pharmaceutical Statistics, 2020. DOI: 10.1002/pst.2028.

[2] Chen J, Hou Y, Chen Z. Statistical inference methods for cumulative incidence function curves at a fixed point in time. Communications in Statistics - Simulation and Computation, 2020, 49(1): 79-94.
}
